% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armadillo_get_number_of_omp_threads}
\alias{armadillo_get_number_of_omp_threads}
\alias{armadillo_set_number_of_omp_threads}
\title{Report (or Set) Maximum Number of OpenMP Threads}
\usage{
armadillo_get_number_of_omp_threads()

armadillo_set_number_of_omp_threads(n)
}
\arguments{
\item{n}{Number of threads to be set}
}
\value{
For the getter, and on a system with OpenMP, the maximum
number of threads that OpenMP may be using and on systems without it,
one. The setter does not return a value.
}
\description{
Report (or Set) Maximum Number of OpenMP Threads
}
