\name{RcppArmadillo-package}
\alias{RcppArmadillo-package}
\alias{RcppArmadillo}
\docType{package}
\title{
Rcpp and Armadillon translation layer
}
\description{
The package eases the integration of Armadillo 
types with Rcpp
}
\details{
\tabular{ll}{
Package: \tab RcppArmadillo\cr
Type: \tab Package\cr
Version: \tab 0.0.0\cr
Date: \tab 2010-02-16\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maintainer: Dirk and Romain <RomainAndDirk@r-enthusiasts.com>
}
\references{
Armadillo project: \url{http://arma.sourceforge.net/}
}
\keyword{ package }

